address cppcheck warnings

+++ examples/c_example.c
@@ -108,15 +108,15 @@ int main(int argc, const char *argv[])
 
 	lcore_id = pcm_getcpu();
 	printf("C:%llu I:%llu, IPC:%3.2f\n",
-		PCM.pcm_c_get_cycles(lcore_id),
-		PCM.pcm_c_get_instr(lcore_id),
+		(unsigned long long)PCM.pcm_c_get_cycles(lcore_id),
+		(unsigned long long)PCM.pcm_c_get_instr(lcore_id),
 		(double)PCM.pcm_c_get_instr(lcore_id)/PCM.pcm_c_get_cycles(lcore_id));
 	printf("CPU%d E0: %llu, E1: %llu, E2: %llu, E3: %llu\n",
 		lcore_id,
-		PCM.pcm_c_get_core_event(lcore_id,0),
-		PCM.pcm_c_get_core_event(lcore_id,1),
-		PCM.pcm_c_get_core_event(lcore_id,2),
-		PCM.pcm_c_get_core_event(lcore_id,3));
+		(unsigned long long)PCM.pcm_c_get_core_event(lcore_id,0),
+		(unsigned long long)PCM.pcm_c_get_core_event(lcore_id,1),
+		(unsigned long long)PCM.pcm_c_get_core_event(lcore_id,2),
+		(unsigned long long)PCM.pcm_c_get_core_event(lcore_id,3));
 
 	return 0;
 }
